[bov] Move cli tests into commands package, like mycoind
