+++ examples/bov/commands/docs.go
@@ -1,0 +1,8 @@
+/*
+Package commands has integration tests of the mycoind
+application.
+
+Tests the application plays together properly with the
+cli commands.
+*/
+package commands
