+++ examples/bov/commands/init_test.go
@@ -1,4 +1,4 @@
-package main_test
+package commands
 
 import (
 	"encoding/json"
