+++ examples/bov/commands/testdata/priv_validator.json
